requirements.txt
Django~=2.0.6
djangorestframework
markdown
django-filter
django-cors-headers